requirements.txt
yfinance
psutil
litellm
python-dotenv
SQLAlchemy
google-cloud-aiplatform
google-cloud-storage
google-genai
gitpython
google-adk